implement authorization policy

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,10 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => 'index','show' ]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -68,11 +72,15 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        if (\Gate::allows('update-question', $question)) {
+        // authorization using Gate
+        // if (\Gate::allows('update-question', $question)) {
 
-            return view("questions.edit",compact('question'));
-        }
-        abort(403, 'Access denied');
+        //     return view("questions.edit",compact('question'));
+        // }
+        // abort(403, 'Access denied');
+
+        $this->authorize('update', $question);
+        return view("questions.edit",compact('question'));
 
     }
 
@@ -85,6 +93,8 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        $this->authorize('update', $question);
+
         $question->update($request->only('title','body'));
 
         return redirect('/questions')->with('success', 'Your Question Has Been Updated');
@@ -98,6 +108,8 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        $this->authorize('delete', $question);
+
         $question->delete();
 
         return redirect('/questions')->with('success','Your Questions Has Been Deleted.');
